+++ EasyFinance.Domain/Financial/Expense.cs
@@ -17,15 +17,15 @@ namespace EasyFinance.Domain.Models.Financial
             User createdBy = default,
             ICollection<Attachment> attachments = default,
             ICollection<ExpenseItem> items = default,
-            decimal Goal = default)
+            int Goal = default)
             : base(name, date, amount, createdBy, attachments, items)
         {
             this.SetGoal(Goal);
         }
 
-        public decimal Goal { get; private set; }
+        public int Goal { get; private set; }
 
-        public void SetGoal(decimal goal)
+        public void SetGoal(int goal)
         {
             if (goal < 0)
                 throw new ValidationException(nameof(this.Goal), string.Format(ValidationMessages.PropertyCantBeLessThanZero, nameof(this.Goal)));
